Added basic CSV datafile support.
git-svn-id: file:///home/yevgeni/dev/repos/srvf/trunk@20 6678053c-22da-4495-bf9c-98dc6e5a3f48

+++ cpp/tests/matrix_tests.cc
@@ -31,8 +31,8 @@ BOOST_AUTO_TEST_CASE(ctor_test3)
   BOOST_CHECK_EQUAL(A.size(),1500);
   BOOST_CHECK(A.data() != (double*)0);
   // check entries using 2-argument accessor
-  for (int i=0; i<A.rows(); ++i){
-    for (int j=0; j<A.cols(); ++j){
+  for (size_t i=0; i<A.rows(); ++i){
+    for (size_t j=0; j<A.cols(); ++j){
       BOOST_CHECK_EQUAL(A(i,j),0.5);
     }
   }
@@ -44,14 +44,14 @@ BOOST_AUTO_TEST_CASE(ctor_test4)
     0.0, 1.0, 2.0, 3.0,
     4.0, 5.0, 6.0, 7.0
   };
-  int rows=2;
-  int cols=4;
-  int sz=rows*cols;
+  size_t rows=2;
+  size_t cols=4;
+  size_t sz=rows*cols;
   srvf::Matrix A(rows, cols, Adata);
   BOOST_CHECK_EQUAL(A.rows(),rows);
   BOOST_CHECK_EQUAL(A.cols(),cols);
   BOOST_CHECK(A.data()!=Adata); // deep copy
-  for (int i=0; i<sz; ++i)
+  for (size_t i=0; i<sz; ++i)
   {
     BOOST_CHECK_EQUAL(Adata[i],A(i));
   }
@@ -64,9 +64,9 @@ BOOST_AUTO_TEST_CASE(ctor_test5)
   BOOST_CHECK_EQUAL(A.rows(),B.rows());
   BOOST_CHECK_EQUAL(A.cols(),B.cols());
   BOOST_CHECK(A.data()!=B.data());
-  for (int i=0; i<A.rows(); ++i)
+  for (size_t i=0; i<A.rows(); ++i)
   {
-    for (int j=0; j<A.cols(); ++j)
+    for (size_t j=0; j<A.cols(); ++j)
     {
       BOOST_CHECK_EQUAL(A(i,j), B(i,j));
     }
@@ -82,7 +82,7 @@ BOOST_AUTO_TEST_CASE(assignment_test1)
   BOOST_CHECK_EQUAL(A.cols(),B.cols());
   BOOST_CHECK_EQUAL(A.size(),B.size());
   BOOST_CHECK(A.data()!=B.data());  // deep copy
-  for (int i=0; i<A.size(); ++i)
+  for (size_t i=0; i<A.size(); ++i)
   {
     BOOST_CHECK_EQUAL(A(i),B(i));
   }
@@ -98,7 +98,7 @@ BOOST_AUTO_TEST_CASE(assignment_test2)
   BOOST_CHECK_EQUAL(A.cols(),20);
   BOOST_CHECK_EQUAL(A.size(),60);
   BOOST_CHECK(A.data()==old_data); // self-assignment shouldn't reallocate data
-  for (int i=0; i<A.size(); ++i)
+  for (size_t i=0; i<A.size(); ++i)
   {
     BOOST_CHECK_EQUAL(A(i),0.5);
   }
@@ -132,9 +132,9 @@ BOOST_AUTO_TEST_CASE(resize_test2)
   BOOST_CHECK_EQUAL(A.cols(),250);
   BOOST_CHECK_EQUAL(A.size(),2500);
   BOOST_CHECK(A.data()!=(double*)0);
-  for (int i=0; i<5; ++i)
+  for (size_t i=0; i<5; ++i)
   {
-    for (int j=0; j<250; ++j)
+    for (size_t j=0; j<250; ++j)
     {
       BOOST_CHECK_EQUAL(A(i,j),0.25);
     }
@@ -149,9 +149,9 @@ BOOST_AUTO_TEST_CASE(resize_test3)
   BOOST_CHECK_EQUAL(A.cols(),10);
   BOOST_CHECK_EQUAL(A.size(),2500);
   BOOST_CHECK(A.data()!=(double*)0);
-  for (int i=0; i<250; ++i)
+  for (size_t i=0; i<250; ++i)
   {
-    for (int j=0; j<5; ++j)
+    for (size_t j=0; j<5; ++j)
     {
       BOOST_CHECK_EQUAL(A(i,j),0.25);
     }
@@ -162,13 +162,13 @@ BOOST_AUTO_TEST_CASE(inplace_plus_test1)
 {
   srvf::Matrix A(1,20);
   srvf::Matrix B(1,20);
-  for (int i=0; i<20; ++i)
+  for (size_t i=0; i<20; ++i)
   {
     A(i)=(double)i;
-    B(i)=(double)(-i);
+    B(i)=-(double)i;
   }
   A += B;
-  for (int i=0; i<20; ++i)
+  for (size_t i=0; i<20; ++i)
   {
     BOOST_CHECK_EQUAL(A(i),0.0);
   }
@@ -179,7 +179,7 @@ BOOST_AUTO_TEST_CASE(inplace_minus_test1)
   srvf::Matrix A(50,200,16.0);
   srvf::Matrix B(50,200,8.0);
   A -= B;
-  for (int i=0; i<A.size(); ++i)
+  for (size_t i=0; i<A.size(); ++i)
   {
     BOOST_CHECK_EQUAL(A(i),8.0);
   }
@@ -189,12 +189,12 @@ BOOST_AUTO_TEST_CASE(inplace_minus_test1)
 BOOST_AUTO_TEST_CASE(inplace_times_test1)
 {
   srvf::Matrix A(1,20);
-  for (int i=0; i<20; ++i)
+  for (size_t i=0; i<20; ++i)
   {
     A(i)=(double)(i);
   }
   A *= A;
-  for (int i=0; i<20; ++i)
+  for (size_t i=0; i<20; ++i)
   {
     BOOST_CHECK_EQUAL(A(i),(double)(i*i));
   }
@@ -206,7 +206,7 @@ BOOST_AUTO_TEST_CASE(inplace_times_test2)
   srvf::Matrix A(100,20,2.0);
   srvf::Matrix B(100,20,0.25);
   A *= B;
-  for (int i=0; i<A.size(); ++i)
+  for (size_t i=0; i<A.size(); ++i)
   {
     BOOST_CHECK_EQUAL(A(i),0.5);
   }
@@ -217,7 +217,7 @@ BOOST_AUTO_TEST_CASE(inplace_div_test1)
 {
   srvf::Matrix A(1,20,16.0);
   A /= A;
-  for (int i=0; i<20; ++i)
+  for (size_t i=0; i<20; ++i)
   {
     BOOST_CHECK_EQUAL(A(i),1.0);
   }
@@ -231,9 +231,9 @@ BOOST_AUTO_TEST_CASE(inplace_div_test2)
   B(5,5)=0.0;
   A /= B;
   BOOST_CHECK(isinf(A(5,5)));
-  for (int i=0; i<10; ++i)
+  for (size_t i=0; i<10; ++i)
   {
-    for (int j=0; j<10; ++j)
+    for (size_t j=0; j<10; ++j)
     {
       if (i!=5 || j!=5)
       {
@@ -247,7 +247,7 @@ BOOST_AUTO_TEST_CASE(inplace_scalar_plus_test1)
 {
   srvf::Matrix A(3,4,0.0);
   A+=1.0;
-  for (int i=0; i<A.size(); ++i)
+  for (size_t i=0; i<A.size(); ++i)
   {
     BOOST_CHECK_EQUAL(A(i),1.0);
   }
@@ -257,7 +257,7 @@ BOOST_AUTO_TEST_CASE(inplace_scalar_minus_test1)
 {
   srvf::Matrix A(3,4,0.0);
   A-=1.0;
-  for (int i=0; i<A.size(); ++i)
+  for (size_t i=0; i<A.size(); ++i)
   {
     BOOST_CHECK_EQUAL(A(i),-1.0);
   }
@@ -267,7 +267,7 @@ BOOST_AUTO_TEST_CASE(inplace_scalar_times_test1)
 {
   srvf::Matrix A(3,4,1.0);
   A*=0.25;
-  for (int i=0; i<A.size(); ++i)
+  for (size_t i=0; i<A.size(); ++i)
   {
     BOOST_CHECK_EQUAL(A(i),0.25);
   }
@@ -277,7 +277,7 @@ BOOST_AUTO_TEST_CASE(inplace_scalar_div_test1)
 {
   srvf::Matrix A(3,4,1.0);
   A/=0.25;
-  for (int i=0; i<A.size(); ++i)
+  for (size_t i=0; i<A.size(); ++i)
   {
     BOOST_CHECK_EQUAL(A(i),4.0);
   }
@@ -287,7 +287,7 @@ BOOST_AUTO_TEST_CASE(scalar_plus_test1)
 {
   srvf::Matrix A(20,32,0.25);
   srvf::Matrix B=A+0.75;
-  for (int i=0; i<B.size(); ++i)
+  for (size_t i=0; i<B.size(); ++i)
   {
     BOOST_CHECK_EQUAL(B(i),1.0);
   }
@@ -297,7 +297,7 @@ BOOST_AUTO_TEST_CASE(scalar_minus_test1)
 {
   srvf::Matrix A(20,32,0.25);
   srvf::Matrix B=A-0.75;
-  for (int i=0; i<B.size(); ++i)
+  for (size_t i=0; i<B.size(); ++i)
   {
     BOOST_CHECK_EQUAL(B(i),-0.5);
   }
@@ -307,7 +307,7 @@ BOOST_AUTO_TEST_CASE(scalar_times_test1)
 {
   srvf::Matrix A(20,32,1.0);
   srvf::Matrix B=A*0.5;
-  for (int i=0; i<B.size(); ++i)
+  for (size_t i=0; i<B.size(); ++i)
   {
     BOOST_CHECK_EQUAL(B(i),0.5);
   }
@@ -317,7 +317,7 @@ BOOST_AUTO_TEST_CASE(scalar_div_test1)
 {
   srvf::Matrix A(20,32,1.0);
   srvf::Matrix B=A/0.5;
-  for (int i=0; i<B.size(); ++i)
+  for (size_t i=0; i<B.size(); ++i)
   {
     BOOST_CHECK_EQUAL(B(i),2.0);
   }
@@ -327,7 +327,7 @@ BOOST_AUTO_TEST_CASE(matmul_test1)
 {
   srvf::Matrix A(2,2,1.0);
   srvf::Matrix A2 = srvf::product(A,A);
-  for (int i=0; i<4; ++i)
+  for (size_t i=0; i<4; ++i)
   {
     BOOST_CHECK_EQUAL(A2(i),2.0);
   }
@@ -336,9 +336,9 @@ BOOST_AUTO_TEST_CASE(matmul_test1)
 BOOST_AUTO_TEST_CASE(transpose_test1)
 {
   srvf::Matrix A(2,3);
-  for (int i=0; i<2; ++i)
+  for (size_t i=0; i<2; ++i)
   {
-    for (int j=0; j<3; ++j)
+    for (size_t j=0; j<3; ++j)
     {
       A(i,j)=3*i+j;
     }
@@ -346,9 +346,9 @@ BOOST_AUTO_TEST_CASE(transpose_test1)
   srvf::Matrix At=srvf::transpose(A);
   BOOST_CHECK_EQUAL(At.rows(),A.cols());
   BOOST_CHECK_EQUAL(At.cols(),A.rows());
-  for (int i=0; i<At.rows(); ++i)
+  for (size_t i=0; i<At.rows(); ++i)
   {
-    for (int j=0; j<At.cols(); ++j)
+    for (size_t j=0; j<At.cols(); ++j)
     {
       BOOST_CHECK_EQUAL(At(i,j),A(j,i));
     }
